backend/requirements.txt
# FastAPI Backend
fastapi>=0.125.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.5
pydantic-settings>=2.6.0
slowapi>=0.1.9
scalar-fastapi>=1.0.0
python-dotenv>=1.0.1
supabase>=2.9.0
redis>=5.2.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
sentry-sdk[fastapi]>=2.48.0
structlog>=24.4.0

# AI/ML Infrastructure
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langsmith>=0.2.0
google-generativeai>=0.8.0

# ETL (PySpark)
pyspark>=3.5.0
delta-spark>=3.0.0
psycopg2-binary>=2.9.10
